[SYCL][Graph][E2E] Remove float usage from Graph tests (#11051) (#16069)
- Removes usage of float from graph E2E tests to avoid precision
differences between host and some GPUs

Addresses issue reported in #10860

(cherry picked from commit 5781e72de67d2bf165caf45cb3ebb60ac7145df0)

+++ sycl/test-e2e/Graph/Inputs/dotp_buffer_reduction.cpp
@@ -8,12 +8,12 @@ int main() {
 
   exp_ext::command_graph Graph{Queue.get_context(), Queue.get_device()};
 
-  float DotpData = 0.f;
+  int DotpData = 0;
 
   const size_t N = 10;
-  std::vector<float> XData(N);
-  std::vector<float> YData(N);
-  std::vector<float> ZData(N);
+  std::vector<int> XData(N);
+  std::vector<int> YData(N);
+  std::vector<int> ZData(N);
 
   {
     buffer DotpBuf(&DotpData, range<1>(1));
@@ -31,9 +31,9 @@ int main() {
       auto Y = YBuf.get_access(CGH);
       auto Z = ZBuf.get_access(CGH);
       CGH.parallel_for(N, [=](id<1> it) {
-        X[it] = 1.0f;
-        Y[it] = 2.0f;
-        Z[it] = 3.0f;
+        X[it] = 1;
+        Y[it] = 2;
+        Z[it] = 3;
       });
     });
 
@@ -65,8 +65,7 @@ int main() {
           auto Dotp = DotpBuf.get_access(CGH);
           auto X = XBuf.get_access(CGH);
           auto Z = ZBuf.get_access(CGH);
-          CGH.parallel_for(range<1>{N},
-                           reduction(DotpBuf, CGH, 0.0f, std::plus()),
+          CGH.parallel_for(range<1>{N}, reduction(DotpBuf, CGH, 0, std::plus()),
                            [=](id<1> it, auto &Sum) { Sum += X[it] * Z[it]; });
         },
         NodeA, NodeB);
